Before quitting, ask for confirmation if there are pending changes
When we're quitting, we want to ask for confirmation first, if we have pending
changes, rather than silently - and quickly - lose them. To do this, we catch
the `before-quit` event in the main process, and unless we were already told to
quit, we'll send a message to the renderer, and stop the quit process.

The renderer, upon receiving the `app.quit-requested` event, can decide how to
proceed: either send an `app.confirm-quit` message back, so the main process
will quit immediately, or send an `app.stop-quit` message, handle any
confirmation necessary, and then restart the quit process as needed.

The main process does set a one second timer before sending the
`app.should-quit?` message, in case the renderer is crashed or otherwise
unresponsive. When the timer is up, we'll quit. Sending an `app.stop-quit`
message back clears the timer.

On the renderer side, we listen to these events in the Header, where the context
bar handling and the confirmation dialog is implemented in the first place. If
we have no context bar visible, we'll quit. If we have one, we'll pop up the
dialog, and set a flag to quit after if the dialog is confirmed.

In practice, all this results in Chrysalis popping up the usual confirm losing
changes dialog when quitting, if there are pending changes, and preventing exit
until confirmed. If there are no pending changes, we quit as usual.

Fixes #917.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>